More layered mappings tests
